Proyecto_IA/apps/backend/pyproject.toml
[project]
name = "proyecto-ia-backend"
version = "0.1.0"
description = "Backend para clasificación jerárquica con SAM-3."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0",
    "pillow>=10.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
